libs-shared/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sahool-shared"
version = "9.0.0"
description = "Sahool Yemen Shared Library - مكتبة سهول اليمن المشتركة"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sahool Yemen Team", email = "dev@sahool.ye"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["sahool", "yemen", "agriculture", "fastapi", "microservices"]

dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "geoalchemy2>=0.14.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/sahool-yemen/sahool-shared"
Documentation = "https://docs.sahool.ye"
Repository = "https://github.com/sahool-yemen/sahool-shared"

[tool.setuptools.packages.find]
where = ["."]
include = ["sahool_shared*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
